prebacen get taxes u portfolio controller
