+++ stock-service/src/main/java/rs/raf/stock_service/controller/PortfolioController.java
@@ -7,11 +7,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import rs.raf.stock_service.domain.dto.PortfolioEntryDto;
+import rs.raf.stock_service.domain.dto.TaxGetResponseDto;
 import rs.raf.stock_service.service.PortfolioService;
 import rs.raf.stock_service.utils.JwtTokenUtil;
 
@@ -46,5 +44,14 @@ public class PortfolioController {
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
         }
+    }
+    @PreAuthorize("hasRole('AGENT') or hasRole('CLIENT')")
+    @GetMapping("/tax/{userId}")
+    @Operation(summary = "Get taxes", description = "Returns paid tax for current year, and unpaid for current month.")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Taxes obtained successfully"),
+    })
+    public ResponseEntity<TaxGetResponseDto>getTaxes(@PathVariable Long userId){
+        return ResponseEntity.ok().body(portfolioService.getTaxes(userId));
     }
 }
